qa-automation/pyproject.toml
[project]
name = "qa-automation"
version = "0.1.0"
description = "Automated Quality Assurance Workflow System"
requires-python = ">=3.11"
authors = [
    {name = "AI Robotics Team"}
]
readme = "README.md"
license = {text = "MIT"}

dependencies = [
    "playwright>=1.40.0",
    "httpx>=0.25.0",
    "fastapi>=0.115.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0.1",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "rich>=13.7.0",
    "click>=8.1.0",
    "jinja2>=3.1.2",
    "jsonschema>=4.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "mypy>=1.7.0",
    "ruff>=0.1.8",
]

[project.scripts]
qa-automation = "qa_automation.cli.main:cli"

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
qa_automation = ["templates/*.html", "templates/*.css"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
